cmd: add custom testnet flags (#2715)
Adds custom testnet flags to `charon create cluster` and `charon run` commands.

category: feature
ticket: #2690

+++ eth2util/eip712/eip712_test.go
@@ -18,7 +18,7 @@ func TestCreatorHash(t *testing.T) {
 		Domain: eip712.Domain{
 			Name:    "Obol",
 			Version: "1",
-			ChainID: uint64(eth2util.Sepolia.ChainID),
+			ChainID: eth2util.Sepolia.ChainID,
 		},
 		Type: eip712.Type{
 			Name: "CreatorConfigHash",
